requirements.txt
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
environs==9.5.0
fastapi==0.103.2
fastapi_cors==0.0.6
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
marshmallow==3.20.1
packaging==23.2
passlib==1.7.4
psutil==5.9.6
psycopg2==2.9.9
pyasn1==0.5.0
pydantic==1.10.9
PyJWT==2.7.0
PySimpleGUI==4.60.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
speedtest-cli==2.1.3
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
websockets==11.0.3
